% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineExperiments.R
\name{combineExperiments}
\alias{combineExperiments}
\title{Combines all of the downloaded datasets into a single summarized experiment object}
\usage{
combineExperiments(datasets)
}
\arguments{
\item{datasets}{}
}
\value{
A summarized experiment object
}
\description{
This is useful for applications such as the TBSignatureProfiler, where a summarized experiment is necessary
}
\examples{
\dontrun{
  combinedExperiments <- combineExperiments(datasets)
}

}
