% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{run_curatedTBExplorer}
\alias{run_curatedTBExplorer}
\title{Run curatedTBExplorer app}
\usage{
run_curatedTBExplorer(dev = FALSE)
}
\arguments{
\item{dev}{Run the application in developer mode}
}
\value{
The shiny app will open
}
\description{
Run curatedTBExplorer app
}
\examples{
\dontrun{
  run_curatedTBExplorer()
}

}
