% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSE.R
\name{toSE}
\alias{toSE}
\title{Converts MultiAssayExperiment to SummarizedExperiment Object}
\usage{
toSE(MAE)
}
\arguments{
\item{MAE}{MultiAssayExperiment}
}
\value{
MAE will be converted to SE
}
\description{
Converts MultiAssayExperiment to SummarizedExperiment Object
}
\examples{
\dontrun{
  SE <- toSE(MAE)
}

}
