% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_genenames.R
\name{update_genenames}
\alias{update_genenames}
\title{Renames Genes Names In a List}
\usage{
update_genenames(siglist)
}
\arguments{
\item{siglist}{A character List}
}
\value{
A character list of adjusted gene names
}
\description{
Renames Genes Names In a List
}
\examples{
\dontrun{
  tt <- c("SEPT4", "FCGR1A", "7-Sep")
  updatedtt <- update_genenames(tt)
}

}
