% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTBsigProfilerFunction.R
\name{runTBsigProfilerFunction}
\alias{runTBsigProfilerFunction}
\title{Runs the TBSignatureProfiler using selected datasets, profiles, and assays}
\usage{
runTBsigProfilerFunction(
  selected_dataset,
  selected_profiles,
  selected_assay,
  selected_algorithm
)
}
\arguments{
\item{selected_dataset}{The selected dataset (summarized experiment) to be viewed. Required.}

\item{selected_profiles}{The group of selected profiles to be viewed. Required.}

\item{selected_assay}{The assay used in the profiler. Required.}

\item{selected_algorithm}{The algorithm to be used in the tbSigProfiler. Required}
}
\value{
A datatable with results
}
\description{
The runTBsigProfilerFunction allows for different datasets, profiles, and assays
to be pulled into a datatable, allowing for easy searching, sorting, and visualization
}
\examples{
\dontrun{
runTBsigProfilerFunction(SE, profiles, assay, algorithm)
}

}
