% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTBsigProfilerFunction.R
\name{runTBsigProfilerFunction}
\alias{runTBsigProfilerFunction}
\title{Runs the TBSignatureProfiler using selected datasets, profiles, and assays}
\usage{
runTBsigProfilerFunction(selected_dataset, selected_profiles)
}
\arguments{
\item{selected_dataset}{The selected dataset (summarized experiment) to be viewed. Required.}

\item{selected_profiles}{The group of selected profiles to be viewed. Required.}
}
\value{
A datatable with results
}
\description{
The runTBsigProfilerFunction allows for different datasets, profiles, and assays
to be pulled into a datatable, allowing for easy searching, sorting, and visualization
}
\examples{
\dontrun{
runTBsigProfilerFunction(SE, profiles)
}

}
